pyproject.toml
[project]
name = "funding-rate-arbitrage-bot"
version = "0.1.0"
description = "A highly automated funding rate arbitrage bot for OKX, featuring a triple-yield engine, active risk management, and Telegram notifications."
authors = [
    {name = "Your Name", email = "you@example.com"},
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "ccxt>=4.4.91",
    "asyncpg>=0.29.0",
    "apscheduler>=3.10.0",
    "python-telegram-bot>=21.0.0",
    "pandas>=2.2.0",
    "PyYAML>=6.0.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "pydantic>=2.7.1",
    "pydantic_settings>=2.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]